Bump go version in `WORKSPACE` file.
This is to enable merging Dependabot created #141. We need to merge both
of these on the same day to ensure that the bazel build does not break.

Unfortunately, Dependabot does not know to update this at the same time
as creating #141 (and similar).

See also #128 which was needed to fix the builds when #121 landed.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>