fix: retries in e2e-bootstrap.sh
This fixes retries in e2e-bootstrap.sh. The reason for previous failure
to retry was due to the fact that bash ignores the `-e` option when
executing a function inside an `if` conditional.
See: https://www.gnu.org/software/bash/manual/html_node/The-Set-Builtin.html

Instead we need to set `+e` before executing the function and set `-e`
inside a sub-shell inside the function. This causes the the function to
terminate early if any command fails, but not terminate the overall
script when the function fails.

We also need to call the function directly and test the `$?` return code
rather than using an `if` conditional so that the `-e` is respected.

Fixes #283

Signed-off-by: Ian Lewis <ianlewis@google.com>